% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hru.R
\name{extract_hru}
\alias{extract_hru}
\title{Extract the hrus table required for topHRU()}
\usage{
extract_hru(file_path)
}
\arguments{
\item{file_path}{character string providing the absolute file path. The
file to read must be either the project.mdb data base file or a .csv
file to where the hrus table from the project.mdb data base was saved.}
}
\value{
Returns the hru_table as a data.frame
}
\description{
Extract the hrus table required for topHRU()
}

