% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pareto.R
\name{plot_pareto}
\alias{plot_pareto}
\title{Plot results of HRU analysis}
\usage{
plot_pareto(eval_hru, area_thrs = NULL, hru_thrs = NULL,
  interactive = TRUE)
}
\arguments{
\item{eval_hru}{list structure that is the result of the function
evaluate_hru()}

\item{area_thrs}{adds a horizontal separation in the visualization at respective threshold value}

\item{hru_thrs}{adds a vertical separation in the visualization at respective threshold value}

\item{interactive}{Logic: \code{TRUE}: Output will be an
interactive plotly object helpful in selecting adequate threshold
combinations, or \code{FALSE} for a static ggplot object useful for
publication reasons.}
}
\value{
Returns an interactive or a static plot object of the aREA
  measure over the resulting number of HRUs of HRUs for the respective
  threshold combinations.
}
\description{
Plot results of HRU analysis
}
\examples{
hru_eval <- evaluate_hru(hru_demo)
plot_pareto(hru_eval, area_thrs = 0.2, interactive = TRUE)
plot_pareto(hru_eval, hru_thrs = 1500, interactive = FALSE)
}

