% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aREA_main.R
\name{topHRU}
\alias{topHRU}
\title{topHRU - threshold optimization for HRUs in SWAT}
\usage{
topHRU(hru_data, luse_thrs = c(0, 20, 5), soil_thrs = c(0, 20, 5),
  slp_thrs = c(0, 20, 5), thrs_type = c("P", "A"), weight = c(1, 1, 1))
}
\arguments{
\item{hru_data}{data.frame with the unaggregated information for the HRUs}

\item{luse_thrs}{Land use threshold vector with min, max and stepwidth.}

\item{soil_thrs}{Soil threshold vector with min, max and stepwidth.}

\item{slp_thrs}{Slope threshold vector with min, max and stepwidth.}

\item{thrs_type}{Threshold type. All threshold vectors must be given as
percentages when set as 'P' or area in ha when set to 'A'.}

\item{weight}{Weights for land use, soil and slope for the calculation of
aREA. Any given numbers will be normalized.}
}
\value{
\code{.$result_all}: data.frame with the complete results of the
  analysis.\cr \code{.$result_nondominated}: Results for the non dominated threshold
  combinations.\cr \code{.$pareto_plot}: Interactive plot of the dominated and non
  dominated threshold combination.
}
\description{
topHRU - threshold optimization for HRUs in SWAT
}
\examples{

hru_data <- hru_data
hru_analysis <- topHRU(hru_data)
View(hru_analysis$result_nondominated)
}

