% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pareto.R
\name{plot_pareto}
\alias{plot_pareto}
\title{Plot results of HRU analysis}
\usage{
plot_pareto(hru_analysis, interactive = TRUE)
}
\arguments{
\item{hru_analysis}{list structure that is the result of the function
topHRU()}

\item{interactive}{Logic: \code{TRUE}: Output will be an
interactive plotly object helpful in selecting adequate threshold
combinations, or \code{FALSE} for a static ggplot object useful for
publication reasons.}
}
\value{
Returns an interactive or a static plot object of the aREA
  measure over the resulting number of HRUs of HRUs for the respective
  threshold combinations.
}
\description{
Plot results of HRU analysis
}

